Refactor packages in modules and core

+++ src/main/java/com/yawarSoft/Core/Config/filter/JWTTokenValidator.java
@@ -68,7 +68,7 @@ public class JWTTokenValidator extends OncePerRequestFilter {
                 String userName = jwtUtils.extractUserName(decodedJWT);
                 //obtenemos authorities en un string separado por comas, esto es por como lo hemos programado
                 String authoritiesString = jwtUtils.getSpecificClaim(decodedJWT, "authorities").asString();
-                Long userId = jwtUtils.getSpecificClaim(decodedJWT, "id").asLong();
+                Integer userId = jwtUtils.getSpecificClaim(decodedJWT, "id").asInt();
 
                 //Se usa AuthorityUtils de spring security que devuelve un collection de Authorities a partir del string separado por comas
                 Collection<? extends GrantedAuthority> authorities = AuthorityUtils.commaSeparatedStringToAuthorityList(authoritiesString);
@@ -93,7 +93,7 @@ public class JWTTokenValidator extends OncePerRequestFilter {
 
     }
 
-    private static SecurityContext setSecurityContext(Long userId, String userName, Collection<? extends GrantedAuthority> authorities) {
+    private static SecurityContext setSecurityContext(Integer userId, String userName, Collection<? extends GrantedAuthority> authorities) {
         SecurityContext context = SecurityContextHolder.getContext();
 
         // Valores por defecto ya que el JWT no los tiene
